Refactor chatbot interface.
